[Apps.] Regenerated demo for hub_msd_sam_a5d27_som1_ek
MH3-93393

+++ apps/hub_msd/firmware/src/config/sam_9x75_curiosity/system/fs/sys_fs.h
@@ -661,7 +661,7 @@ typedef void (* SYS_FS_EVENT_HANDLER)
 typedef struct 
 {
     /* File size */
-    uint64_t    fsize;
+    uint32_t    fsize;
     /* Last modified date */
     uint16_t    fdate;
     /* Last modified time */
